Extract multiple_use_keys into a class property.
